refactor(ui): extract feed/trace.ts — TraceCandidateGroup for inspector panel
